backend/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.34.0
google-genai==1.64.0
pydantic==2.12.5
motor==3.7.1
pymongo==4.16.0
python-dotenv==1.2.1
